requirements.txt
fastapi>=0.110
aiosqlite>=0.20
asyncpg>=0.29
google-adk
greenlet>=3.0
httpx>=0.26
kafka-python>=2.0
neo4j>=5.19
opentelemetry-api>=1.23
opentelemetry-sdk>=1.23
prometheus-client>=0.20
psycopg[binary]>=3.1
pydantic>=2.6
python-dotenv>=1.0
pytest>=7.4
pyyaml>=6.0
redis>=5.0
requests>=2.31
ruff>=0.4
uvicorn[standard]>=0.27
